medic: move https support to own script, add medic service
This adds a simple systemd wrapper for Medic's 'yarn start' to run the service at server startup.

Also, while testing deployment for medic I hit a limit on the number of LetsEncrypt calls we can make.  For now I'm factoring out the call to set up the domain cert to a separate script.  We can still connect via ssh tunneling.