Update aws-api-setup for external database
Previously, aws-api-setup would initialize a web server and database on the same node.  Now, since we have an RDS database with regular backup snapshots, this modifies aws-api-setup/api-init to use the existing database.  PostgreSQL has the concept of a '~/.pgpass' file that contains credentials for one or more databases.  This script reuses that file (and expects it to already exist) to gather information about the database instance 'db0', and initializes the web server's '.env' to point there.