RequireMFA: organize code to simplify auditing group membership
The definitions of policies have a lot of details, but the important part that will likely change over time is defining who is in which groups.  To facilitate this, I now define each group and all its policies together, and then gather all the group memberships together at the end of the file so they are easy to find.

Also, I still define aws_iam_user blocks for each user referenced to validate they exist, but now I'm just using userids directly in the group membership blocks since it is much easier to read and sanity check.