terraform: don't specify S3 key
When I set up the S3 bucket encryption, I used the id of the existing key I saw.  The result was S3 trying to use a customer-managed key with a description that indicates it is for RDS.  Encryption fails.

Since I just want to use the default key anyway, I can avoid specifying the key at all and the right thing should happen.

It turned out that kms created a customer-managed key