RDSLogs: successfully archived logs
Firing an empty test event to the lambda successfully copied the staging-non-pii database logs to an S3 bucket under the correct path.  stdout from the lambda execution successfully ended up in a CloudWatch log.

This enables a CloudWatch event on an interval to trigger the lambda.  I'll test that by checking later that logs are continuing to be uploaded.

This archives RDS logs for the staging database