RDSLogs: increase timeout to 10 minutes (#354)
It may only matter for the first couple uploads, but with the 30-second timeout there were cases where the job was timing out.  At the beginning it has to upload all the existing log files, and this can take a while.  With 10 minutes, everything seems to have been successfully archived.