Add CLI support for editing part of a visit's JSON
This adds a command "edit" to the visit-mod script.  The "edit" command takes a (pii|nonpii), a row id/csruid-prefix, and a JSON path.  It opens $EDITOR, $VISUAL, or falls back to vim on the specified JSON sub-object.  After saving, it invokes colordiff on the before and after and asks if you want to commit.  If you say yes, the change is written back to the db.