Clean up terraform TODOs
The main change here is to enable deletion protection and final snapshots for the databases.  I had these off while I was testing because I needed to destroy and recreate the databases repeatedly.  Now that there is real data, these should be on.

The provisioning scripts generally log every command to a readable /setup.log.  I had marked these "TODO remove" because sometimes the logs contain credentials.  However, this shouldn't matter on a machine when the machine is transient and does not exist in steady state.  Since the logs are so useful in cases where something goes wrong, and in those cases the credentials don't matter anyway (since we regenerate on the final successful run), I'm keeping the logs and just removing the TODO comments.

The final change here is to avoid re-paving dev servers when the init script changes.  If there are ever changes that matter for security, we can announce the need to make the change, but generally developers are responsible for their machines.