Introduce email into welcomeNext screen (#482)
* Introduce email into welcomeNext screen

A design change was requested to frontload email getting in order to reduce risk of customer disappointment at the last step.  We create common strings, introduce an email field to the welcomeNext screen, and write the email to the store.  The giftcard screen will then have the email prefilled, and is already authored to handle that case.

* Updated per code review feedback;  all requested changes made.