Updated all product strings to latest (#492)
* Updated all product strings to latest

This was a huge pain, because the original spreadsheet had dozens of rows rearranged.  Took 4+ hours to get this back into reasonable shape.  Unfortunately, not 100% of the original orders were preserved, but I hope that's ok for now.  I don't personally think it's worth spending more time on this.

* Updated all product strings to latest

This was a huge pain, because the original spreadsheet had dozens of rows rearranged.  Took 4+ hours to get this back into reasonable shape.  Unfortunately, not 100% of the original orders were preserved, but I hope that's ok for now.  I don't personally think it's worth spending more time on this.

* Significant manual shuffle to restore diffability

* Fix comma issue -- oops