FEV-40: upload photos separately
The survey document is regularly edited, saved, and uploaded as the user moves through the survey.  Adding a base64-encoded jpeg image into the survey documents results in repeatedly doing this process on a ~10 MB photo blob instead of a ~10 KB blob.

These commits move photos into a separate document that is uploaded once, and then we record the csruid of the photo in the survey document.

To facilitate knowing the csruid of the photo so we can link it into the survey, these changes update DocumentUploader to take a csruid as input, and update the UI layer to generate a new csruid every time we clear state.