Make our barcode entry less ambiguous (#507)
* Make our barcode entry less ambiguous

Anticipated user issues are around the barcode sticker having a "KIT" prefix.  We don't want them to type it in.  So we've done a few things:

1. We now specifically extract 8-char hex codes
2. We force-prefix the user's text entry with "KIT ", so that what they type conforms to what our blue sticker says.
3. We normalize all references to "barcode" and "8-character code".
4. Constants and stuff.

* Remove push notification screen

* Make our barcode entry less ambiguous

Anticipated user issues are around the barcode sticker having a "KIT" prefix.  We don't want them to type it in.  So we've done a few things:

1. We now specifically extract 8-char hex codes
2. We force-prefix the user's text entry with "KIT ", so that what they type conforms to what our blue sticker says.
3. We normalize all references to "barcode" and "8-character code".
4. Constants and stuff.