FEV-145: Fix 'remove test strip' sequence (#526)
In the case where you complete the survey before the timer finishes, we were sending you to the FinishStrip screen without first sending you to the TestStripReady screen.  So the flow didn't make sense.  Whereas our code was already doing this correctly in cases where you take longer on the survey than 10 mins.  This corrects that error.