FEV-210 Updated MonthPicker to get right default (#544)
MonthPicker used to not let the parent component know if the user left the default date selected (in our case, it's the last date given).  This would mean if you got a flu shot in the same month as the current date, the parent would think you didn't set a date at all.  We now notify the parent after the first render of the component so that the parent has the right date even if you Ok the picker without ever changing away from the default value.