FEV-149: Timer doesn't count down when app minimized (#541)
* FEV-149: Timer doesn't count down when app minimized

Original code relied on setTimeout when the app was in the foreground, but didn't count time down when app is backgrounded.  Its implementation was also subject to performance-based drift, because it relied on sixty timely setTimeout events in order to be right.  We now map time to the wall clock so that we can solve both issues.

Verified this works both when you let the time count down the whole way, and if you demo-mode-skip the timer.  With and without app minimization.

Cleaned up the timeout handle in cases where we unmount (not entirely a big deal, but we weren't using the original _timer at all;  must have had it in order to do this very thing).  Rounded, to prevent jitter/drift based on system load.

* Timer hygiene

Cleaned up timer when done, and also going with ceil!