FEV-200 Account for removed gift page (#536)
* FEV-200 Account for removed gift page

We've now removed the final gift card page, and so we need to add a reminder on the Thanks screen that you're indeed getting your gift card soon.  This adds that string, and populates it with your email.

* Updated to include reduxWriter, as required by linting.

It worked even without this, but it's not clean that way.

* Typesafety fix.

* Updated date promise to give more leeway

We're not positive we'll really turn things around in 24 hours.  This gives us more room.

* FEV-200 Account for removed gift page

We've now removed the final gift card page, and so we need to add a reminder on the Thanks screen that you're indeed getting your gift card soon.  This adds that string, and populates it with your email.

* Updated to include reduxWriter, as required by linting.

It worked even without this, but it's not clean that way.

* Typesafety fix.

* Updated date promise to give more leeway

We're not positive we'll really turn things around in 24 hours.  This gives us more room.

* Final string tweaks and corrections

* Flu@Home -> flu@home