FEV-170 Implement multiple exclusiveOption possibilities (#542)
We have at least two survey questions now for which multiple items are exclusive.  This extends exclusiveOption to support that possibility (and I left its grammar to default to the singular, since that's most common).

Tested the Medical Conditions and Insurance questions to make sure they work, and then also the Race question to verify that classic single-exclusives still work.