FEV-237 Updated Study Info screen to reduce confusion (#557)
One of the reasons we were rejected from the App Store was specifically the Study Info screen.  Reviewers thought that since the screen only refers to BBI/UW, it's misleading to the public to say the publisher is Audere.  We're now making this screen more explicitly show that Audere owns and operates the app (and is thus rightfully its publisher), and that BBI/UW specifically are working just on the study within.