FEV-216 Implement markdown links (#563)
* FEV-216 Implement markdown links

This handles mailto, which is the only scenario we currently have.  But instead of custom hacking that, I instead implemented the standard markdown syntax for all links:  [read me](http://bbc.co.uk).

I did try to import linkifyjs in order to do reliable auto-detection of all links in text, but that only works on Web and React right now, and doesn't build correctly in our codebase.  And it felt over-the-top heavy for what we need.  Implementing markdown syntax has the happy side effect of also supporting link text that's independent of URL.

* Introduce stronger typing.

Not sure this'll work, but CircleCi's detection of type issues is different from what VS Code shows me for the same file.  The original commit is clean in VS Code.

* Updated per feedback on PR #562 (thanks, @rightparen!).  Also fixed merge error.