Fix build break in prod (#574)
Prod linking step couldn't find libRNFirebase.a, yet Debug could.

Do I know what caused this?  Yes.  A dangling reference in our xcodeproj, which is not a straight-up link to the right library.

Do I know why removing and re-adding the library fixed it?  No.  And it bothers me a bit.  But we need to get a build out.

If you know the black magic, please enlighten.  Thanks!