Update camera screen for test strip (#586)
* Update camera screen for test strip

After some back and forth, I think the consensus is to try to test this method of getting people to take the right photo.  The previous overlay made it hard to focus/see the actual strip you're photographing.  This one attempts to put it on the left, and changes the wording in hopes that it's clearer.

I'm still not super-happy with this, TBH.  But after much back and forth with Shawna and David, I think this is the best we have at the moment.

* Optimized the CSS to be alphabetized and to remove outdated opacity.