FEV-68 Implement Firebase nav tracking (#567)
* FEV-68 Implement Firebase nav tracking

Now that Firebase is in, this gets really easy.  In this diff:

1. We initialize Firebase to track user metrics
2. We intercept navigation to log forward/backward and drawer events
3. We set Xcode to log to Firebug's debug console whenever you run Debug.  This reduces our tainting of real user data.

At some point, when we feel good enough about all of this, we could likely get rid of the timestamp-based event logger we also introduced a while back.  Would get rid of FEV-224.  And Firebase tools are really good for visualizing/tracking these events.

* Removed needless imports

* Updated to move setupJest into test dir