FEV-210 Allow user to cancel out of date selection (#603)
Without this change, if you bring up the date picker (e.g. flu shot date) and cancel it, you still get the current month as your selected date.  In other words, once you bring up the picker, you MUST pick a date... and if you don't, we'll pick the current month.

I've changed this to allow for the case where you bring the picker up but cancel the picker. In those cases, we leave the date as unset.  Note, however, that because RN's Picker doesn't support the concept of an "unselected" state once you've selected something, there's no way for us to cleanly support the case where you select something, click Done on the picker, and then want to open up the date to clear it because you've changed your mind.

TBH, I'm not sure all this effort/trouble is worth it to support the default case.  But Sam made a good point that it's odd, once you bring up the dialog, that you currently can't change your mind;  cancelling, without this change, leaves the current month selected.