Implement device detection / exclusion from our usage logs (#607)
* Implement device detection / exclusion from our usage logs

In order to not taint our production logging data, this introduces three abilities:

1. Filtering by IP
2. Filtering by device name
3. Overriding filtering (for engineers debugging firebase logging)

In order to implement this, we've added react-native-device-info via `yarn add` and `react-native link`.  In addition, we did the classic build-hacks addition.  I've verified that the build succeeds in both debug and prod, and also that `yarn test` succeeds.  I've also verified that the Firebase console shows logging statements when I set ignoreTrackingExclusionAndTrackMe = true, and that without the setting, WeWork IPs and my phone's device name are indeed excluded.

* Fix build breaks and type issue

Somehowe I managed to `yarn add` on my own machine but missed package.json and yarn.lock to the build. Le sigh.  Sorry about that.

* Simplified tracking to focus only on demo & production as inputs

* Updated await strategy

* Re-kicking Circle CI in attempt to reproduce snyk issue.