Install Crashlytics for native crashes (#650)
Though native crashes are hardly likely or interesting for us, this is a required step prior to integrating React Native's Crashlytics integration (which I hope will give us meaning RN crash / error logs).

Most of this diff is just integrating the native SDK.  I removed the old Fabric data from our plist, which was added during ejection, because it's unclear to me that we used any Fabric services prior to this (since this diff's the one that actually added Fabric to our Pods).  Leaving that in there confuses Crashlytics (which integrated Fabric's old users) as to which API key you're actually using.