Install Crashlytics for JS (#653)
* Install Crashlytics for JS

Now that we have native Crashlytics, this adds react-native-fabric for RN-level JS exception reporting.  We parse out the first line of the error message to be its "domain" (Firebase groups errors by this), and then include the rest of the stack in userInfo.

I've tested and logged an intentional JS error (by setting connect's StoreState to null) to see it show up on the Firebase Crashlytics console.  All seems to be working.

* Fix jest for tests