Fix duplicate symbols for Crashlytics (#656)
Don't honestly know how this ever built for me, but it did, including all the way through to getting crash logs on Firebase's server console.  When I looked at the first diff, I thought it was teird that this was doubled up in the Xcode proj, but I didn't want to mess with it since it seemed to be working.  Turns out I should have manually edited the file and taken it out, to prevent double-linking.

I still don't understand how this happened.  Sorry.