Inserts Crashlytics logs for nav and state (#659)
This attempts to store specific checkpoints in the code that then get uploaded alongside crash logs so that we have a better sense what the user was doing before the crash.  Note that logs are only persisted/set for the current session of the user... so, for instance, if the app insta-crashes upon launch, before any nav events, you won't get nav logs from their previous session.