Log middleware with Crashlytics
This ends up being a fair amount of logging overhead - each transition is about 256 bytes.  However, we are seeing state=null when components try to bind to state.  I suspect it might be an issue with one of our middlewares, so this logs on the way in and out for every middleware in the stack.