Fix tracker logging to include production (#681)
Can I be this lame?  Unfortunately, yes.  Did a direct copy/paste from Slack and didn't read carefully.  Also, couldn't tell from Firebase Console because prior versions were still logging to us, and we've only now begun to taper because of adoption of build 0.2.1+.  And in debugging this, I set the override flag, so never tested isProduction to be true.  Le sigh.

Anyway.  We kinda need to ship this, I think.