FEV-311 Make email verification independent of component references (#677)
* FEV-311 Make email verification independent of component references

Email validation used to require an instance of the EmailInput component.  But callers sometimes wanted to validate emails _before_ that component's reference even exists (e.g. call _canProceed earlier in the rendering stack than a ref can be gotten).

This pulls validation out of the component, making it possible to proceed on screens where this circular requirement previously made it not work.

* Moved isValidEmail to util/check