FEV-332 Log app version with consent shown (#698)
* FEV-332 Log app version with consent shown

In order to keep our customer promise (while adjusting gift card amounts), we need to be able to know which dollar amount we told them they were getting in the app. We could in theory sniff the consent, but that doesn't localize very well and is just a bit hokey. Michael and I thought it'd be better to instead write out an app version alongside the consent text so that our backend can then map the app version into the right payment when sending the email list to the gift card folks.

* Updated to use hash, and to process the lib/ correctly.