FEV-332 Insert build number into consent data (#712)
We need to track the $ we promised a user, so that app upgrades which change the default $ don't end up reneging on a previous promise.

This changes us from writing a hash into writing a build number, so that backend code can be written more like `if (buildNum > 47) { price = $25 }` rather than having to update the backend with every single hash of builds we release that now contain a different price.

Long term, we'd likely want to write the actual price into this.  But previous discussions with Michael had us leaning away from doing that for now (out of concerns of hacking).