FEV-332 Remove gift amount references in Part Deux (#719)
There's a chance that between when a participant completes Part 1 and Part 2, we've changed our mminds about the amount that the gift card is for.  We've taken care of all of this on the backend and in our database (to keep the customer promise), but we now should remove all specific dollar amount references in Part 2.

Change is fairly straightforward:  we remove the amounts in Part 2, leaving references only in Part 1's Welcome and Consent.