Various UI fixes (#725)
We were already handling button widths in ButtonGrid... but we left one question sizing itself for button widths, which led to a squeeze in button sizes.  We now use ButtonGrid's default with, which looks much better.

I had a lame boolean logic fail in RadioGrid.  Fixed that.

Turned the youngChildren question into a radio, which makes much more sense.  OptionLists that are exclusive should probably be turned into radios.  I went through the product, and I think this is the only one.