FEV-386 Force lowercase barcodes (#745)
All our barcodes are lowercase.  This forces it, because we've had some backend code matching issues due to title casing and what not.

This takes a different approach to how ManualEntryScreen works -- it now instead normalizes/forces lowercase on inbound props, and maintains this.state.barcode* always in lowercase.  So we can tighten the RE that checks for barcodes, and reduce the places where we call .toLowerCase().