FEV-408 Throttle remoteConfig calls (#748)
Firebase throttles remote config loads on the client side if you call more than ~5 times an hour (the exact figure is never exposed by Firebase staff... internet lore arrived at this figure).

We now cache settings for an hour.  What this means is that, when we ship a change, it'll take reloading the app up to an hour later from the last time you loaded the app before you'll get the change.  However, for all new users and for users who last launched the app over an hour ago, the change will be available immediately for them.

Should not affect development, since we `enableDeveloperMode` in debug, and that avoids all caching.