FEV-398 dbcli: trim spaces from csruid arguments
Sometimes a csruid begins with a '-'.

I haven't been able to convince yargs to parse a positional parameter to a sub-command that begins with '-'.  I tried putting a bare '--' before the argument, and I tried calling it with --row=-...  Neither works.

As a workaround, I am trimming spaces so I can quote the csruid with a leading space: " -..."  Hopefully this should enable me to pass these in.