FEV-297 use --maxWorkers=1 with test-watch target
FeverConsentMailerEndpoint tries to generate emails for all valid surveys in fever_current_surveys.  If we run multiple tests concurrently, other tests will generate (and delete) surveys.  If a survey exists when the mailer queries for it and is deleted when the mailer goes to add a record that mail was sent, we get a foreign key constraint violation.

Most of the test targets were already using --maxWorkers=1.  This makes them all uniform.  It also aligns the text to make it easy to see visually that they are consistent and where they differ.