Temporary fix for survey batch
Kit ordering is currently broken because every request after the first one fails.  We are updating the model every time we process a batch, and evidently Sequelize models' `hasMany` is not idempotent.

This does the brain-dead fix of setting a local boolean to make sure we only call it once.  We should probably be doing this explicitly in the models file somehow.