Add funnel logging for app backgrounding and results (#799)
We want a better understanding of when/how people are quitting out of the app (e.g. when they fall off during the 10-min timer), so we add logging for app foregrounding/backgrounding.  We also want a faster pipeline to figuring out our potential flu positives/negatives, so we now log events for user-reported reads of the strip.