Boot in Android (#784)
* Boot the product in Android!

Good times here.

* Simplifying to the essentials to get running.

This takes out Crashlytics and Remote Config to the point where the app runs with as minimal changes as possible. Created FEV-433 to track.

* Yet more simplifications down to the basics of what's needed

* Boot the product in Android!

Good times here.

* Simplifying to the essentials to get running.

This takes out Crashlytics and Remote Config to the point where the app runs with as minimal changes as possible. Created FEV-433 to track.

* Yet more simplifications down to the basics of what's needed

* Attempted to upgrade to Expo v32

This follows the web instructions on upgrading from Expo 31 to 32, and yet doesn't run when it boots.  Currently red-screens, claiming that Expo isn't running.  Preliminary research on the web hasn't yielded a believable answer to what's wrong.  (I've tried upgrading/downgrading expo-cli, for example, as several developers claim, but it hasn't solved the problem).

In this diff, I also removed api_key, since it probably shouldn't be committed to the codebase.

* Updating expo dependencies

* Add images for Android