CircleCI: Snyk scan only if this branch changes deps
We were comparing local deps to current master deps.  If master has changed deps since my branch's base commit, we would run a Snyk scan even if this branch did not change deps.

This compares the branch deps with the merge-base deps where we split off instead so we only run if this branch changes deps.