Track CSRUID into Firebase for positivity correlation (#879)
We currently don't have a way to track per-channel positivity (i.e. link lab results to the ad channel that led to the user).  This is a significant shortcoming of our current logging plan.

This writes CSRUID into Firebase via an event whenever the CSRUID is established:  whether upon first launch or any subsequent launch.  The reason we write on subsequent launches is in case DeviceID (which Firebase uses to identify the "same" user) may not be available by the time initializeCSRUID() finishes.  And since the latter is async, it's hard to guarantee which finishes first.