FEV-519: Add firestore capability to Sniffles (#920)
* FEV-519: Add firestore capability to Sniffles

This enables Sniffles to commit data to Firebase Cloud Firestore.  We `yarn add firebase` and then create a utility file to do the basic functions.

Everything try-catches out of an abundance of caution.

Tested on my machine using the real apiKey.  Will pull the real key from CircleCI in a subsequent diff.

* Added firebase key access