FEV-438: Enable overriding kit block (#927)
* FEV-438: Enable overriding kit block

This enables testing with a long-press on any menu image (the same area as demo-mode's triple-tap).  If you long-press on the image, you'll alter the kit-blocking status locally.  This is completely independent of demo mode (i.e. this will work even in prod).  If you'd like this to only be possible in demo mode, LMK.

* Post-merge fixup