FluApi: notify Slack when ELB gets 5XX
This moves the encryption key used for SSM parameters to global, and references a new shared SSM parameter that puts the Slack hook URL is a shared location.  It creates a shared modules/notifier and calls it from new flu/2-notifier-{staging,prod} modules.  The SNS topic defined there is then shared with api and lambda so we can notify from anywhere above the db.  (and there are no dependencies there, so if desired later we could swap the order of db and notifier to allow the db to notify).

Finally, this adds a sample usage of the SNS topic to create alarms for ELB/ApplicationELB when they observe a 5XX error code from something.  This might eventually be too noisy, but starting off verbose to see what we hit.