PLAT-11: Get Crashlytics to work, fix notification build break (#1031)
Turns out our google-services.json was missing a few critical things to talk with Crashlytics, and was instead still using the default Expo app settings.  Fixed that now.

Master build was also broken by react-native-push-notification, which was asking for a version of firebase-messages (18.0.0) which hasn't even been publicly released.  We now lock them to the same version as we're already pulling in (17.6.0).