terraform: notifier: fix npm install
I had problems running the `npm install` in terraform/modules/notifier/init-local on OSX, but that may be because I use MacPorts.  It ran fine on a Linux VM.  On OSX it resulted in the everything in the directory having attributes like "d?????????" for directories and "?????????" for files.

This reverts an attempted fix that didn't work.  The correct fix was to run on Linux.