FirebaseReceiver: address feedback from another PR + tweaks
This fixes `updates()` to use the sender's timestamp to order updates.  It shouldn't matter, but this could conceivably make things more robust if we ever saw document corruption that caused failures.  If that happened, we would always get the 256 most recent updates.  If that window ever got clogged up with corrupt documents that caused failures, this query allows us to move on and still read more recent documents while we debug the corrupt ones.

After fixing that, running the integration test triggers an error because we need new indexes.  Since we have to create the indexes anyway, I'm taking the opportunity to fix some naming issues.