Store data in Firestore (#1087)
* Store data in Firestore

This is the first commit to begin storing our data into Firestore.  We've left pouch in place for now (PLAT-50), and will remove it only once we're sure this is all working.

We assume there's no PII in the survey.  This is an issue to be resolved by ongoing discussions with our partners (particularly, whether the HREC will approve us not having an option to email the survey to the participant).  If we introduce PII, we'll parse it out and encrypt it prior to upload.

This enables offline caching for Firestore.  It assumes we'll save the images separately from the doc... but ideally, we'd save the image in the doc itself (if the whole thing is under 1MB) because then Firestore will automatically cache locally and upload the next time you're connected.

* Updated allowed field values for lastWritten

* Type edits, and removed calls to Fever uploader endpoint

* Also fixed FluTrack

* Unwind docId rename to affect only Cough

* Updates from code review