Remove AnalyticsBatcher (#1118)
* Remove AnalyticsBatcher

We've transitioned to using Firebase for analytics as well as Crashlytics for logging, and so coughProtocol can likely be simplified by removing AnalyticsInfo as an upload type.  This cascades all the consequences.

One thing:  LMK your opinion on whether we should keep a bunch of debug logging, in which case I might recommend that we instead implement a debug logger such that engineers can selectively turn on logging levels and see the logs come through the console.  As it stands, it's not clear to me how much we actively use those logs, vs. simply adding console.log deliberately when debugging specific issues.

* Logging updates and also crashlytics verification