Change CSRUID to GUID+Installation
We originally intended CSRUIDs to be cryptographically secure as part of the security model to prevent someone from modifying an existing record.  However, this does nothing to prevent someone from contributing bogus data as new records.  Since determining that, we have been going with GUIDs as "good enough".  This makes that official, and also augments the UID with a suffix that correlates items that come from a single device.