firebase: add photo protocol
This adds both a sender and receiver that use Firestore to index photos uploaded to Firebase Storage.  It follows the same pattern we use for surveys to mark buffered items that have not yet been handled by the receiver.

It is possible to see (harmless) transient failures since the client does not wait for photo upload to complete before enqueuing an entry for the photo and writing the photo id into the survey.  In the race, the server may notice and try to download a photo that is not completely uploaded.  This should fail, the item will not be marked as read, and the server will retry later.  If this race ends up being common (or if it doesn't fail for some reason and we notice missing photos) we may need to work out a way to make this more robust.

Also, I feel bad about using undocumented api, but I did not see a documented way to get the project id / default bucket name from a firebase app or firebase storage.  However, if this ever causes failures, the failures are easily visible in the logs and the server just stops importing, but we don't lose any state.