Remove DocumentUploader
We plan to use Firebase instead of DocumentUploader to upload surveys and photos, so this is no longer needed.

Some of my upcoming Firestore/FirebaseStorage changes modify the schema in a way that breaks DocumentUploader, so I am removing it here rather than updating unused code there.

The only part that is still useful is the storage encryption password logic.  I just moved that function from DocumentUploader to store/index.