FEV-588 Remove Branch and Firebase Dynamic Links (#1175)
* FEV-588 Remove Branch and Firebase Dynamic Links

We don't track app installations now, and will not likely use advertising in the future to push the app.  We never quite got attribution tracking working 100% anyway, so getting rid of the extra native binaries for Branch and Firebase Dynamic Links is probably a good idea.

Removes all related calls, both on the native and in the RN side.  Booted and navigated on both platforms.

* Merged podfile.lock