Reverse proxy: retry on transient forwarding failures
There has been a problem with the ssh reverse proxy from sunflower.  I suspect there are cases where a retry fails to forward the port because the previous iteration has not yet fully cleaned up.  In that case, we have a successful connection but no forwarded port, and this condition persists indefinitely.

This makes the whole connection fail if we cannot forward the port, since that is the whole point of the connection.  That way failing to forward will fall back on the retry logic to reconnect.  If the port issue is transient, this should fix it.

I also discovered the `-N` option in ssh so we don't need the sleep command.