Address GitHub-reported vulnerabilities
Three new vulnerability alerts came up today:

1. sequelize-cli in FluApi/yarn.lock (moderate severity)

There is no current fix for this, but I think it's okay.  The issue is that passwords with odd characters may escape encoding and result in passwords being displayed.  However, this is only under devDependencies for developer use on a trusted local laptop.  WONTFIX

2. handlebars in FluApi/yarn.lock (high severity)

This is fixed in handlebars 4.1.2, we have a direct dependency, and we use this on the web portal.  The current commit upgrades this to include the fix.

3. handlebars in FluLambda/yarn.lock (high severity)

In FluLambda, we only have an indirect dependency via devDependencies.  In this case, handlebars appears to be used by jest to generate test reports.  However, we don't use this feature of jest, so this is not very interesting.  WONTFIX