FEV-640 Rolling window for flu vaccinations (#1229)
Our medical partners in Fever insisted on Jan 2018 as a start date for vaccinations.  Now that we're done with that, we'd like to move to a rolling window of vaccinations so that we always show up to the same month a year before (because the question says "in the last year").  There are probably arguments that this might be relevant beyond (e.g. if you open the app in Dec 2019, but got a flushot in Oct 2018), but for now, this seems reasonable.