Fix tests after deps upgrade
For FluApi and FluLambda, this fixes warnings in the tests.

For FluTrack and FluStudy_au, this reverts upgrading `jest-expo`.  We were upgrading this because its version of `handlebars` has a security issue.  However, we don't use that functionality in `jest-expo`, and upgrading `jest-expo` requires other upgrades that have file format changes.

...except that js-yaml also comes in from jest-expo and react-native.  So I'll enter a separate issue to upgrade these.