FluApi: fix portal login
I had tried to consolidate hash functions, not realizing that there were subtle differences.  The original portal login just does an update with each string in turn.  The consolidated function joins the strings with a space and hashes the resulting string.

We are moving the portal eventually to use a more secure hash.  Until then, put the original hash function back.

Tested by verifying I could see the same failure, and with this fix I can log in with the old password again.