FEV-700 All your surveyButton are belong to us (#1301)
We introduced <DropDown>, riding on top of the existing SurveyQuestion infrastructure, but didn't quite adhere to the full way that answerOptions are managed in the uploader.  There are several places in our codebase that understand the notion of surveyButtons, and treat them like options.  Specifically, we assume that all SurveyQuestions that have buttons as options for answers use text keys of "surveyButton:myTextKey".  DropDown didn't obey this convention, and so the labels being uploaded for it were instead myTextKey (which is what `t()` defaults to if it can't find your key).

We now conform DropDown to how the rest of the codebase treats option buttons for survey questions:

1. We move its choices into surveyButton
2. We put DropDown and DropDownModal into the surveyButton translation domain