PLAT-120 FluApi: add SqlLock and ensure only one cough importer
We hit an issue where ~100 photos were uploaded in an hour.  It took a while for the photos to transfer, and AWS Lambda decided to start another call even though the 300-second timeout had not yet expired.  We ended up with 3 calls all trying to transfer the same photos.  This logged the photos to the db audit log several times and used a lot of space.  There are multiple issues to fix here, but one is that we should not run multiple concurrent importers.

This adds the ability to use a discretionary lock in the db to ensure that only one handler is running at a time.