PLAT-128: avoid sending headers after body
This time with more testing...

It turns out we don't need `setTimeout()` if we are sending progress bytes, so I removed that.

It also turns out that `res.json()` tries to set the "Content-Type" header if it hasn't already been set, and then calls `res.send()`, which tries to set "Content-Length" as well.  So this manages the whole process, calling `res.type("json")`, writing out the body, and finally calling `res.end()`.

This seemed to work locally.