FEV-673 Update defective strip test (#1418)
This adds the right cautionary/helpful paragraph right before the URL.  It was either a regression from a previous change, or was never implemented.

Deliberately decided not to share out the almost-identical paragraph with the positive and negatiev screens, because:  a) they're not identical, and b) sharing the same text with the positive instance would require writing a new component like ScreenText that concatenates two translated strings, or alternatively changing ScreenText to accept non-label literal text.  Both of these seem overkill to me, especially in the AUS case where we're unlikely to translate at all.